fix installation

+++ install.sh
@@ -6,7 +6,9 @@ install -d /usr/local/share/rasplay2
 install *py requirements.txt /usr/local/share/rasplay2/
 install -d /usr/local/etc/rasplay2
 install config.yaml.dist /usr/local/etc/rasplay2/config.yaml.dist
-install config.yaml.dist /usr/local/etc/rasplay2/config.yaml
+if ! [ -f /usr/local/etc/rasplay2/config.yaml ]; then
+   install config.yaml.dist /usr/local/etc/rasplay2/config.yaml
+fi
 python3 -m venv /usr/local/share/rasplay2/venv
 /usr/local/share/rasplay2/venv/bin/python3 -m pip install -r /usr/local/share/rasplay2/requirements.txt
 install -m 644 rasplay2.service /etc/systemd/system/
@@ -30,4 +32,8 @@ rasplay2 ALL=(ALL) NOPASSWD: /sbin/shutdown, /sbin/reboot, /sbin/systemctl ^rest
 EOL
 
 systemctl daemon-reload
+if systemctl is-active rasplay2.service ; then
+   systemctl restart rasplay2.service
+fi
 systemctl status rasplay2.service
+
